Create shclasses submodule that contains the files SHCoeffs, SHGrid, and SHWindow
