+++ www/pyshcoeffs.html
@@ -41,12 +41,12 @@
 		</tr>
 	
 		<tr>
-		<td class="left"><tt>SHRealCoefficients</tt></td> 
+		<td class="left"><tt>SHRealCoeffs</tt></td> 
 		<td class="right">Real spherical harmonic coefficient class.</td>
 		</tr>
 		
 		<tr>
-		<td class="left"><tt>SHComplexCoefficients</tt></td> 
+		<td class="left"><tt>SHComplexCoeffs</tt></td> 
 		<td class="right">Complex spherical harmonic coefficient class.</td>
 		</tr>
