+++ www/pyshwindow.html
@@ -42,12 +42,12 @@
 		</tr>
 	
 		<tr>
-		<td class="left"><tt>SymmetricWindow</tt></td> 
+		<td class="left"><tt>SHSymmetricWindow</tt></td> 
 		<td class="right">Class for windows concentrated within a spherical cap.</td>
 		</tr>
 		
 		<tr>
-		<td class="left"><tt>AsymmetricWindow</tt></td> 
+		<td class="left"><tt>SHAsymmetricWindow</tt></td> 
 		<td class="right">Class for windows concentrated in an arbitrary domain.</td>
 		</tr>
